Module icu_provider_tzif::data::posix

Structs for holding data encoded by POSIX time-zone strings, as specified by https://www.gnu.org/software/libc/manual/html_node/TZ-Variable.html

Structs

DstTransitionInfo

A struct for holding DST transition info.

PosixTzString

A struct for holding data encoded by a POSIX time-zone string.

TransitionDate

A struct to hold a DST transition date.

ZoneVariantInfo

A struct to hold a time-zone variant name and its offset. The offset is how many hours must be added to the time to reach UTC.

Enums

TransitionDay

A struct for defining a DST transition day.